Change folder structure and naming
