+++ extension/src/config/ctpConfig.js
@@ -1,1 +1,4 @@
 console.log('configuring the CTP platform')
+
+function init() {
+}
